rag-backend/requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
httpx
python-dotenv 